rename rcrainfo_service.py to rcrainfo_client.py
