+++ server/core/services/__init__.py
@@ -1,4 +1,4 @@
-from .rcrainfo_service import RcraClient, get_rcra_client
+from .rcrainfo_client import RcraClient, get_rcra_client
 from .task_service import TaskService, get_task_status, launch_example_task
 
 __all__ = [
